+++ packages/uql-orm/src/@types/index.d.ts
@@ -1,1 +1,2 @@
+import './bunSql';
 import './jest';